Close assets when not contained in position.balances
There was a misunderstanding in previous commits. It was expected
that position.balances contains all assets that have changed
(for example when balance was set to 0), while in reality it only
contains active, non-zero assets. So we need to figure out ourselves
which previously active assets were closed. This code does this.